requirements.txt
torch
torchaudio
torchvision
numpy
opencv-python
tqdm
Pillow
natsort